Rename the file too
